Minor fixes

+++ test/dcl_tester.py
@@ -22,7 +22,7 @@ class DCLTester:
             for batch_idx, d in enumerate(self.dataloader):
                 inputs, labels = d
                 inputs = inputs.to(self.device)
-                labels = Variable(torch.from_numpy(np.array(labels))).to(self.device)
+                labels = Variable(torch.from_numpy(np.array(labels))).long().to(self.device)
                 cls_outputs, adv_outputs = model(inputs, train=False)
                 outputs = cls_outputs + adv_outputs[:, 0:200] + adv_outputs[:, 200:2 * 200]
                 loss = self.loss(outputs, labels)
